+++ index.html
@@ -11,7 +11,7 @@
   <!-- Подключение файла function.js из директории js -->
   <script src="/js/functions.js"></script>
   <!-- Подключение файла main.js из директории js -->
-  <script src="/js/main.js"></script>
+  <script src="/js/main.js" type="module"></script>
 </head>
 
 <body>
