+++ js/util.js
@@ -1,0 +1,20 @@
+//Функция получающая случайное целое положительное число из диапазона [a,b]
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+const getRandomId = (arrId, a, b) => {
+  let id = getRandomInteger(a, b);
+  while(arrId.includes(id)){
+    id = getRandomInteger(a, b);
+  }
+  arrId.push(id);
+  return id;
+};
+
+export {getRandomArrayElement, getRandomId, getRandomInteger};
